src/app/+app/+proveedores/supplier-modal/supplier-modal.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { SupplierModalComponent } from './supplier-modal.component.ts';

describe('Component: SupplierModal', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [SupplierModalComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([SupplierModalComponent],
      (component: SupplierModalComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(SupplierModalComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(SupplierModalComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <supplier-modal></supplier-modal>
  `,
  directives: [SupplierModalComponent]
})
class SupplierModalComponentTestController {
}

